DataViews: Optimize the patterns dataviews by extracting the fields definition
