+++ packages/edit-site/src/components/page-patterns/style.scss
@@ -1,10 +1,92 @@
+.edit-site-patterns__delete-modal {
+	width: $modal-width-small;
+}
+
+.page-patterns-preview-field {
+	display: flex;
+	flex-direction: column;
+	height: 100%;
+
+	.dataviews-view-table & {
+		width: 96px;
+		flex-grow: 0;
+		border-radius: 2px;
+
+		.page-patterns-preview-field__button {
+			border-radius: 2px;
+		}
+	}
+
+	.page-patterns-preview-field__button {
+		box-shadow: none;
+		border: none;
+		padding: 0;
+		background-color: unset;
+		box-sizing: border-box;
+		cursor: pointer;
+		overflow: hidden;
+		height: 100%;
+		border-radius: $grid-unit-05;
+
+		&:focus-visible {
+			box-shadow: 0 0 0 var(--wp-admin-border-width-focus) var(--wp-admin-theme-color);
+			// Windows High Contrast mode will show this outline, but not the box-shadow.
+			outline: 2px solid transparent;
+		}
+
+		&[aria-disabled="true"] {
+			cursor: default;
+		}
+	}
+}
+
+.edit-site-patterns__pattern-icon {
+	fill: var(--wp-block-synced-color);
+	flex-shrink: 0;
+}
+
+.edit-site-patterns__pattern-lock-icon {
+	min-width: min-content;
+}
+
 .edit-site-patterns__section-header {
+	border-bottom: 1px solid #f0f0f0;
+	padding: $grid-unit-20 $grid-unit-60;
+	position: sticky;
+	top: 0;
+	z-index: 2;
+	flex-shrink: 0;
+	transition: padding ease-out 0.1s;
+	@include reduce-motion("transition");
+	min-height: $grid-unit-50;
+
+	.edit-site-patterns__title {
+		min-height: $grid-unit-50;
+
+		.components-heading {
+			flex-grow: 1;
+			flex-basis: 0;
+			white-space: nowrap;
+		}
+	}
+
+	.edit-site-patterns__sub-title {
+		margin-bottom: $grid-unit-10;
+	}
+
 	.screen-reader-shortcut:focus {
 		top: 0;
 	}
 }
 
 .edit-site-patterns__pattern-title {
+	display: block;
+	width: 100%;
+	white-space: nowrap;
+	overflow: hidden;
+	text-overflow: ellipsis;
+	color: inherit;
+
 	.is-link {
 		text-decoration: none;
 		color: $gray-200;
@@ -26,93 +108,7 @@
 	}
 }
 
-.edit-site-patterns__delete-modal {
-	width: $modal-width-small;
-}
-
 .edit-site-page-patterns-dataviews {
-	.page-patterns-preview-field {
-		display: flex;
-		flex-direction: column;
-		height: 100%;
-
-		&.is-viewtype-table {
-			width: 96px;
-			flex-grow: 0;
-			border-radius: 2px;
-
-			.page-patterns-preview-field__button {
-				border-radius: 2px;
-			}
-		}
-
-		.page-patterns-preview-field__button {
-			box-shadow: none;
-			border: none;
-			padding: 0;
-			background-color: unset;
-			box-sizing: border-box;
-			cursor: pointer;
-			overflow: hidden;
-			height: 100%;
-			border-radius: $grid-unit-05;
-
-			&:focus-visible {
-				box-shadow: 0 0 0 var(--wp-admin-border-width-focus) var(--wp-admin-theme-color);
-				// Windows High Contrast mode will show this outline, but not the box-shadow.
-				outline: 2px solid transparent;
-			}
-
-			&[aria-disabled="true"] {
-				cursor: default;
-			}
-		}
-	}
-
-	.edit-site-patterns__pattern-icon {
-		fill: var(--wp-block-synced-color);
-		flex-shrink: 0;
-	}
-
-	.edit-site-patterns__pattern-lock-icon {
-		min-width: min-content;
-	}
-
-	.edit-site-patterns__section-header {
-		border-bottom: 1px solid #f0f0f0;
-		padding: $grid-unit-20 $grid-unit-60;
-		position: sticky;
-		top: 0;
-		z-index: 2;
-		flex-shrink: 0;
-		transition: padding ease-out 0.1s;
-		@include reduce-motion("transition");
-		min-height: $grid-unit-50;
-
-		.edit-site-patterns__title {
-			min-height: $grid-unit-50;
-
-			.components-heading {
-				flex-grow: 1;
-				flex-basis: 0;
-				white-space: nowrap;
-			}
-		}
-
-		.edit-site-patterns__sub-title {
-			margin-bottom: $grid-unit-10;
-		}
-	}
-
-	.edit-site-patterns__pattern-title {
-		display: block;
-		width: 100%;
-		white-space: nowrap;
-		overflow: hidden;
-		text-overflow: ellipsis;
-		color: inherit;
-	}
-
 	.dataviews-pagination {
 		z-index: z-index(".edit-site-patterns__dataviews-list-pagination");
 	}
